Changed clark_trot_controller to open_loop_controller
