+++ open_loop_controller/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.0.2)
-project(clark_trot_controller)
+project(open_loop_controller)
 
 ## Compile as C++11, supported in ROS Kinetic and newer
 add_compile_options(-std=c++11)
@@ -32,19 +32,19 @@ include_directories(
 )
 
 ## Declare a C++ library for control
-add_library(clark_trot_controller
-	src/clark_trot_controller.cpp
+add_library(open_loop_controller
+	src/open_loop_controller.cpp
 )
-add_dependencies(clark_trot_controller ${clark_trot_controller_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
+add_dependencies(open_loop_controller ${open_loop_controller} ${catkin_EXPORTED_TARGETS})
 
 ## Declare a C++ executable
 ## With catkin_make all packages are built within a single CMake context
 ## The recommended prefix ensures that target names across packages don't collide
-add_executable(clark_trot_controller_node src/clark_trot_controller_node.cpp)
+add_executable(open_loop_controller_node src/open_loop_controller_node.cpp)
 
 ## Specify libraries to link a library or executable target against
-target_link_libraries(clark_trot_controller_node
-  clark_trot_controller
+target_link_libraries(open_loop_controller_node
+  open_loop_controller
 	${catkin_LIBRARIES}
 )
 
@@ -53,7 +53,7 @@ target_link_libraries(clark_trot_controller_node
 #############
 
 ## Google test file for clark_trot_controller library
-catkin_add_gtest(clark_trot_controller_test test/test_clark_trot_controller.cpp)
-target_link_libraries(clark_trot_controller_test ${catkin_LIBRARIES} clark_trot_controller)
+catkin_add_gtest(open_loop_controller_test test/test_open_loop_controller.cpp)
+target_link_libraries(open_loop_controller_test ${catkin_LIBRARIES} open_loop_controller)
