+++ open_loop_controller/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>clark_trot_controller</name>
+  <name>open_loop_controller</name>
   <version>0.0.0</version>
-  <description>The clark_trot_controller package</description>
+  <description>The open_loop_controller package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
   <!-- Example:  -->
