+++ open_loop_controller/src/open_loop_controller_node.cpp
@@ -1,14 +1,14 @@
 #include <ros/ros.h>
 #include <iostream>
 
-#include "clark_trot_controller/clark_trot_controller.h"
+#include "open_loop_controller/open_loop_controller.h"
 
 int main(int argc, char** argv) {
 	ros::init(argc, argv, "controller_node");
 	ros::NodeHandle nh;
 
-	ClarkTrotController clark_trot_controller(nh);
-	clark_trot_controller.spin();
+	OpenLoopController open_loop_controller(nh);
+	open_loop_controller.spin();
 
 	return 0;
 }
