+++ open_loop_controller/test/test_open_loop_controller.cpp
@@ -1,0 +1,17 @@
+#include <ros/ros.h>
+#include <gtest/gtest.h>
+
+#include "open_loop_controller/open_loop_controller.h"
+
+TEST(OpenLoopController, testTrue) {
+	ros::NodeHandle nh;
+	OpenLoopController open_loop_controller(nh);
+	EXPECT_EQ(1 + 1, 2);
+}
+
+int main(int argc, char** argv) {
+	testing::InitGoogleTest(&argc, argv);
+	ros::init(argc, argv, "open_loop_controller_tester");
+
+	return RUN_ALL_TESTS();
+}
